Extract settings to new module
