+++ app/src/main/java/app/tivi/inject/AppModuleBinds.kt
@@ -17,15 +17,12 @@
 package app.tivi.inject
 
 import app.tivi.appinitializers.AppInitializer
-import app.tivi.appinitializers.ArchTaskExecutorInitializer
 import app.tivi.appinitializers.ClearGlideCacheInitializer
 import app.tivi.appinitializers.EmojiInitializer
 import app.tivi.appinitializers.PreferencesInitializer
 import app.tivi.appinitializers.ThreeTenBpInitializer
 import app.tivi.appinitializers.TimberInitializer
 import app.tivi.appinitializers.TmdbInitializer
-import app.tivi.settings.TiviPreferences
-import app.tivi.settings.TiviPreferencesImpl
 import app.tivi.util.AndroidPowerController
 import app.tivi.util.Logger
 import app.tivi.util.PowerController
@@ -47,10 +44,6 @@ abstract class AppModuleBinds {
     @Binds
     abstract fun provideLogger(bind: TiviLogger): Logger
 
-    @Singleton
-    @Binds
-    abstract fun providePreferences(bind: TiviPreferencesImpl): TiviPreferences
-
     @Binds
     @IntoSet
     abstract fun provideEmojiInitializer(bind: EmojiInitializer): AppInitializer
@@ -66,10 +59,6 @@ abstract class AppModuleBinds {
     @Binds
     @IntoSet
     abstract fun providePreferencesInitializer(bind: PreferencesInitializer): AppInitializer
-
-    @Binds
-    @IntoSet
-    abstract fun provideArchTaskExecutorInitializer(bind: ArchTaskExecutorInitializer): AppInitializer
 
     @Binds
     @IntoSet
