+++ settings.gradle
@@ -52,5 +52,6 @@ include ':ui-trending'
 include ':ui-popular'
 include ':ui-recommended'
 include ':ui-search'
+include ':ui-settings'
 include ':ui-account'
 include ':app'
