+++ ui-settings/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!--
-  ~ Copyright 2019 Google LLC
+  ~ Copyright 2020 Google LLC.
   ~
   ~ Licensed under the Apache License, Version 2.0 (the "License");
   ~ you may not use this file except in compliance with the License.
@@ -13,24 +13,10 @@
   ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
   ~ See the License for the specific language governing permissions and
   ~ limitations under the License.
+  ~
   -->
 
-<resources>
-    <id name="image_key" />
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="app.tivi.settings">
 
-    <id name="navigation_discover" />
-    <id name="navigation_trending" />
-    <id name="navigation_popular" />
-    <id name="navigation_followed" />
-    <id name="navigation_watched" />
-    <id name="navigation_recommended" />
-    <id name="navigation_search" />
-    <id name="navigation_settings" />
-    <id name="navigation_licences" />
-    <id name="navigation_show_details" />
-
-    <id name="activity_show_details" />
-
-    <id name="motion_layout_listeners" />
-
-</resources>
+</manifest>
