+++ ui-settings/src/main/java/app/tivi/settings/SettingsInject.kt
@@ -1,0 +1,50 @@
+/*
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package app.tivi.settings
+
+import android.content.Context
+import android.content.SharedPreferences
+import androidx.preference.PreferenceManager
+import dagger.Binds
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ApplicationComponent
+import dagger.hilt.android.qualifiers.ApplicationContext
+import javax.inject.Named
+import javax.inject.Singleton
+
+@InstallIn(ApplicationComponent::class)
+@Module
+internal abstract class SettingsModuleBinds {
+    @Singleton
+    @Binds
+    abstract fun providePreferences(bind: TiviPreferencesImpl): TiviPreferences
+}
+
+@InstallIn(ApplicationComponent::class)
+@Module
+internal object SettingsModule {
+    @Named("app")
+    @Provides
+    @Singleton
+    fun provideAppPreferences(
+        @ApplicationContext context: Context
+    ): SharedPreferences {
+        return PreferenceManager.getDefaultSharedPreferences(context)
+    }
+}
