+++ ui-settings/src/main/res/values/preferences.xml
