+++ ui-settings/src/main/res/xml/preferences.xml
